test/common: Ensure to unload test modules
The _cleanup is easily missed to be called at places where trap
ERR is set, ex security.sh.

The patch moves all the modprobes into the _init in test/common
and sets the trap on exit for "any" of the reasons to invoke
_cleanup.

The patch also gracefully skips the test if the module load fails
instead of continuing, which otherwise makes the tests fail
erroneously in random places during the tests.

Signed-off-by: Shivaprasad G Bhat <sbhat@linux.ibm.com>